--- UACloudLibraryServer/NodeSetIndex/NodeSetModelIndexer.cs
@@ -1,75 +1,0 @@
-/* ========================================================================
- * Copyright (c) 2005-2022 The OPC Foundation, Inc. All rights reserved.
- *
- * OPC Foundation MIT License 1.00
- *
- * Permission is hereby granted, free of charge, to any person
- * obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without
- * restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- * OTHER DEALINGS IN THE SOFTWARE.
- *
- * The complete license agreement can be found here:
- * http://opcfoundation.org/License/MIT/1.00/
- * ======================================================================*/
-
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
-using Opc.Ua.Cloud.Library.Models;
-using Opc.Ua.Cloud.Library.NodeSetIndex;
-using Opc.Ua.Export;
-
-namespace Opc.Ua.Cloud.Library
-{
-    public class NodeSetModelIndexer
-    {
-        private AppDbContext _dbContext;
-        private readonly ILogger _logger;
-
-        public NodeSetModelIndexer(AppDbContext dbContext, ILoggerFactory logger)
-        {
-            _dbContext = dbContext;
-            _logger = logger.CreateLogger("NodeSetModelIndexer");
-        }
-
-        public async Task IndexNodeSetModelAsync(UANodeSet nodeset)
-        {
-            var nodesetModel = await CloudLibNodeSetModel.FromModelAsync(nodeset.Models[0], _dbContext).ConfigureAwait(false);
-
-            await NodeModelFactoryOpc.LoadNodeSetAsync(
-                new CloudLibDbOpcUaContext(_dbContext, _logger, null),
-                nodeset,
-                nodesetModel,
-                new Dictionary<string, string>(),
-                true
-            ).ConfigureAwait(false);
-
-            if (!_dbContext.Set<NodeSetModel>().Any(nsm => nsm.ModelUri == nodesetModel.ModelUri && nsm.PublicationDate == nodesetModel.PublicationDate))
-            {
-                _dbContext.NodeSetsWithUnapproved.Add(nodesetModel);
-            }
-            else
-            {
-                _dbContext.NodeSetsWithUnapproved.Update(nodesetModel);
-            }
-
-            await _dbContext.SaveChangesAsync().ConfigureAwait(false);
-        }
-    }
-}
